requirements.txt
langchain-openai
langchain
streamlit
python-dotenv
aioconsole
asyncio